pyproject.toml
[project]
name = "aegis"
version = "0.1.0"
description = "Resource Governor for Multi-Agent Systems — budget, compute, and state control plane."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Aegis Contributors" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis[hiredis]>=5.2.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "fakeredis>=2.26",
    "httpx",              # for FastAPI TestClient
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
